Membuat Error Handling & Store Route

+++ PangkalanData/database/migrations/2021_01_28_152850_create_anggaran_table.php
@@ -17,7 +17,7 @@ class CreateAnggaranTable extends Migration
             $table->id();
             $table->string('unit');
             $table->string('tahun_anggaran');
-            $table->integer('nilai_anggaran');
+            $table->bigInteger('nilai_anggaran');
             $table->timestamps();
         });
     }
